+++ R/deprecated.R
@@ -1,0 +1,11 @@
+#' Deprecated functions
+#'
+#' See the "Functions" section for replacements
+#' @name deprecated
+#'
+#' @section Functions:
+#'
+#' - `as.root_criterion()`: Use [as_root_criterion()]  instead.
+#'
+#' - `is.root_criterion()`: Use [is_root_criterion()] instead.
+NULL