+++ R/rrmake.R
@@ -5,8 +5,11 @@ make_find_root_file <- function(criterion) {
   }))
 }
 
-make_fix_root_file <- function(criterion, path) {
+make_fix_root_file <- function(criterion, path, subdir=NULL) {
   root <- find_root(criterion = criterion, path = path)
+  if (!is.null(subdir)) {
+    root <- file.path(root, subdir)
+  }
   eval(bquote(function(...) {
     file.path(.(root), ...)
   }))